Strike and Quadrant to Bearing
At times, and particularly when working with a computer, it is easier to visualize directional data using bearing (-90ª - 90°, with 0 as N-S).  I have written two Matlab functions that will make use of either Azimuth data or Quadrant data and convert it into a bearing measure.

I had to do this many times during my PhD, so I hope that someone else may benefit from this package.